+++ cogs/admin.py
@@ -10,7 +10,7 @@ from utils.config import config
 from utils.database import *
 from utils.discord import get_user_avatar
 from utils.general import join_list, log
-from utils.tibia import tibia_worlds, get_character, NetworkError, Character, get_voc_abb_and_emoji
+from utils.tibia import get_character, NetworkError, Character, get_voc_abb_and_emoji
 
 
 class Admin:
@@ -219,89 +219,6 @@ class Admin:
 
         set_server_property("events_channel", ctx.guild.id, channel.id)
         await ctx.send("This server's announcement channel was changed successfully.\n"
-                       "If the channel becomes unavailable for me in any way, I will try to use the highest channel"
-                       " I can see on the list.")
-
-    @commands.command(name="setlevelsdeathschannel", aliases=["setlevelschannel", "setdeathschannel"])
-    @checks.is_admin()
-    @commands.guild_only()
-    @checks.is_not_lite()
-    async def set_levels_deaths_channel(self, ctx: commands.Context, *, name: str = None):
-        """Sets or checks the channel used for level up and deaths.
-
-        if no name is provided, the bot will display the current option.
-        If no channel is set, the bot will use the top channel it can write on."""
-        def check(m):
-            return m.author == ctx.author and m.channel == ctx.channel
-
-        top_channel = self.bot.get_top_channel(ctx.guild, True)
-        current_channel = get_server_property("levels_channel", ctx.guild.id, is_int=True)
-        # Show currently set channel
-        if name is None:
-            if current_channel is None:
-                if top_channel is None:
-                    await ctx.send("This server has no channel set and I can't use any of the channels.")
-                    return
-                await ctx.send(f"This server has no channel set, I use {top_channel.mention} cause it's the highest on "
-                               f"the list I can use.")
-                return
-            channel = ctx.guild.get_channel(int(current_channel))
-            # Channel doesn't exist anymore
-            if channel is None:
-                if top_channel is None:
-                    await ctx.send("The channel previously set doesn't seem to exist anymore and I don't have a "
-                                   "channel to use.")
-                else:
-                    await ctx.send(f"The channel previously set doesn't seem to exist anymore. I'm using "
-                                   f"{top_channel.mention} meanwhile.")
-                return
-            permissions = channel.permissions_for(ctx.me)
-            if not permissions.read_messages or not permissions.send_messages:
-                if top_channel is None:
-                    await ctx.send(f"The channel is {channel.mention}, but I don't have permissions to use it and I "
-                                   f"don't have a channel to use.")
-                else:
-                    await ctx.send(f"The channel is {channel.mention}, but I don't have permissions to use it. "
-                                   f"I'm using {top_channel.mention} meanwhile.")
-                return
-            await ctx.send(f"This server's levels and deaths channel is {channel.mention}.")
-            return
-
-        if name.lower() in ["clear", "none", "delete", "remove"]:
-            await ctx.send("Are you sure you want to delete this server's levels and deaths channel? `yes/no`")
-            try:
-                reply = await self.bot.wait_for("message", timeout=50.0, check=check)
-                if reply.content.lower() not in ["yes", "y"]:
-                    await ctx.send("No changes were made then.")
-                    return
-            except asyncio.TimeoutError:
-                await ctx.send("I guess you changed your mind...")
-                return
-            set_server_property("levels_channel", ctx.guild.id, None)
-            await ctx.send("This server's levels and deaths channel was removed.")
-            return
-        try:
-            channel = await commands.TextChannelConverter().convert(ctx, name)
-        except commands.BadArgument:
-            await ctx.send("I couldn't find that channel.")
-            return
-        permissions = channel.permissions_for(ctx.me)
-        if not permissions.read_messages or not permissions.send_messages:
-            await ctx.send("I don't have permission to use {0.mention}.".format(channel))
-            return
-
-        await ctx.send("Are you sure you want {0.mention} as the level and deaths channel? `yes/no`".format(channel))
-        try:
-            reply = await self.bot.wait_for("message", timeout=120.0, check=check)
-            if reply.content.lower() not in ["yes", "y"]:
-                await ctx.send("No changes were made then.")
-                return
-        except asyncio.TimeoutError:
-            await ctx.send("I guess you changed your mind...")
-            return
-
-        set_server_property("levels_channel", ctx.guild.id, channel.id)
-        await ctx.send("This server's level and deaths channel was changed successfully.\n"
                        "If the channel becomes unavailable for me in any way, I will try to use the highest channel"
                        " I can see on the list.")